Add configuration if not there /etc/ipsec.conf

+++ central/scripts/installer/install_lnx.sh
@@ -40,14 +40,25 @@ subj=$(cat ${home}client_info.md |grep "subj")
 subj=${subj#subj: *}
 
 if [ $(cat /etc/ipsec.conf |grep "$subj" | wc -l) -eq 0 ];then
+
    echo "It looks like that there is no configuration for your subject in /etc/ipsec.conf"
-   echo "Would you like to overwrite the whole /etc/ipsec.conf?"
-   read -p "If you are not connected with an other strongswan Gateway you can use y (yN): " overwrite
-   if [[ $overwrite == "[yY]" ]]; then
-      read -p "Complete /etc/ipsec.conf will be overwritten by merge. Ok? (yN): " overwrite
+   if [ $(cat /etc/ipsec.conf |grep "conn rt.conn" | wc -l) -eq 0 ];then
+      echo "Adding configuration to /etc/ipsec.conf"
+      echo "" >> /etc/ipsec.conf
+      echo "Rothirsch Tech. connection" >> /etc/ipsec.conf
+      sed -n '/conn rt.conn/,$p' ${home}ipsec.conf >> /etc/ipsec.conf
+      ipsec restart
+      ipsec up rt.conn
+   else
+      echo "But there is an existing old connection, would you like to overwrite the whole /etc/ipsec.conf?"
+      read -p "If you are not connected with an other strongswan Gateway you can use y (yN): " overwrite
       if [[ $overwrite == "[yY]" ]]; then
-         mv ${home}ipsec.conf /etc/ipsec.conf
-         ipsec restart
+         read -p "Complete /etc/ipsec.conf will be overwritten by merge. Ok? (yN): " overwrite
+         if [[ $overwrite == "[yY]" ]]; then
+            mv ${home}ipsec.conf /etc/ipsec.conf
+            ipsec restart
+            ipsec up rt.conn
+         fi
       fi
    fi
